optee_test: Lock to a known good version of xtest
This is a known working version of xtest with the current OPTEE git HEAD
for WaRP7. Some additional work may have to be done to rebase OPTEE and
xtest onto an official OPTEE tag aligning user-space and OPTEE space to a
given release. For now this patch gets tip-of-tree OPTEE working with xtest
as produced by yocto.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>