firmware-warp7.bb: change to firmware-imx_%.bbappend
Use a bbappend file rather than a new recipe. This allows us to not
modify the local.conf when including the updated Broadcom firmware into
the build.

This patch rolls in minor changes, including fixing whitespace and
adding COMPATIBLE_MACHINE, which will be needed later when we use our
own mkimage to support building optee features in u-boot.

Signed-off-by: Ying-Chun Liu (PaulLiu) <paulliu@debian.org>