image_sign_mbl: add bitbake class to sign a given binary
This bitbake class provides inheritable routines that can be used to sign
an arbitrary piece of data.

The main routine provided here is
- image_sign_mbl_binary
  Which takes the following inputs
  $1 = Path to binary being signed as it exists in the root filesystem
  $2 = Name of the board (required to resolve CSF path)
  $3 = Input image name
  $4 = Output image name "-signed will be appended to this name"
  $5 = u-boot CONFIG address to find to populate HAB file
  $6 = CSF file - which needs to be populated with "HAB" entries
  $7 = name of u-boot board config
  and produces a signed binary ${WORKDIR}/temp/$4-signed as an output

The signing routine relies on
- A CSF file provisioned by warp7-csf.bb
- A set of Keys provisioned by warp7-keys.bb
- The CST utility provisioned by imx7-cst.bb

Similarly build-time data such as
- u-boot.cfg
- boardname.cfgtmp

is used to derive the load address of the HAB signed binary. The HAB
"Blocks =" directive in the CSF requires the absolute location of a CSF
binary in DRAM when cryptographically signing the binary.

In order to facilitate that very rigid mapping of addresses it is desirable
to have only one place we change addresses. That place is the u-boot
environment and most specifically the #defines associated with the
parameter $5 above.

CONFIG_LOADADDR - being an example of where u-boot wants to load a Linux
                  kernel. As mentioned the HAB input file wants to know
                  that address also - so rather than maintain the address
                  both in BSP recipe and in u-boot config - we just have
                  the bitbake recipe extract the data directly from
                  u-boot.cfg

You must pass a u-boot.cfg a boardname.cfgtmp and a CONFIG_ADDRESSHERE of
some description to sign a binary.

An IMX signed binary contains two core pieces of data
1. An IVT (image vector table) header
2. A CSF (command sequence file) header or footer

The IVT contains an address that is the absolute address of the CSF and the
CSF contains "HAB" blocks which tell the BootROM where to locate keys etc
(using absolute physical addresses).

The uboot mkimage utility if called with the proper parameters will
actually output a "HAB Blocks" string - which we need to capture, reformat
and write to the CSF file, before generating a signed image.

It really is a marathon. Bottom line being if you want to sign a binary you
need to have an address to load that binary pre-allocated in u-boot.cfg.

You can re-use an already allocated address - such as CONFIG_LOADADDR
provided the two things you want to validate at that address don't need to
occupy the same address location at the same time.

CONFIG_LOADADDR is reused for kernel and boot.scr for this reason though a
future set of modification might separate that behaviour out.

This first pass produces the necessary signed binaries when asked in the
right way - as part of yocto.

If you change an address in u-boot all you have to do is launch a yocoto
build and the next produced set of images will automatically be signed with
the correct HAB data, which is all that really matters.

Input for NXP should center around
1. Removing the requirement for a CSF file (please)
2. Removing the dependency on absolute addresses. It should be enough for
   the BootROM to take a pointer to an IVT and then use addressing that is
   relative to the IVT to accomplish its goals.
3. Make the source code for the utility open please. There is a bug with
   CST 2.3.2 around either file names being too long or the token "+"
   appearing as part of the filename, something easily fixed if you have
  the code.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>