u-boot: add compile step to save data for signing
When signing binaries using NXP's signing utility we need to save

u-boot.cfg
boarname.cfgtmp

In addition to sign u-boot itself we cannot use the u-boot.imx created by
the build system - instead we need to re-create the .imx version of u-boot
from the original u-boot.bin but capture the output of mkimage to a file
for further processing.

This patch saves the necessary u-boot data to a temporary directory defined
by a dependency on image_sign_mbl.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>