warp7-keys: add a recipe to retrieve keys for signing an image
We need to retrieve data in a bitbake/yocto friend way in order to produced
signed images via the build system.

This patch assumes an internal github private account which the team will
share access to in order to do signing of images. When handing over MBL to
another team, that team will want to use their own keys for signing.

One way to do that is to have a git repo internally and to point this
recipe to that github repo instead of to the mbl keys, which are being
shared purely for development convenience.

This recipe pulls down the necessary key data from github and places the
data into the native-sdk path.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>