u-boot: add signed image production to build process
This patch adds automatic u-boot binary image signing to the build process.
The CSF file associated with this new signed image is u-boot_sign.csf.

A file called u-boot.imx-signed will appear in the deploy directory and
this file can subsequently be used to burn into an SD card for secure or
non-secure boot depending on if the secure-boot bit is flipped in the OTP
fuses.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>