optee-os: add ability to use private key from CST for TAs
OP-TEE Trusted Applications (TAs) get signed with a decrypted private key.
This signing process let's the run-time OP-TEE software know if an
application is authenticated or not.

This patch makes it possible to define

- WARP7_KEY_NAME_ENCRYPTED
  A path to an encrypted private key

- WARP7_KEY_PASS_FILE
  A path to a key-pass file to decrypt the private key

- WARP7_KEY_NAME_DECRYPTED
  A path to a decrypted target key.

If both WARP7_KEY_NAME_ENCRYPTED and WARP7_KEY_PASS_FILE are defined then
the encrypted key will be decrypted to WARP7_KEY_NAME_DECRYPTED.

If WARP7_KEY_NAME_DECRYPTED is defined on its own then the decryption
operation will be skipped.

If WARP7_KEY_NAME_DECRYPTED is not set then the default OP-TEE private key
will be used.

This first version of the recipe uses one of the private keys created by
the NXP CST tool as the private key source but, any private key can be
used.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>