u-boot: use UART6 as console
Update the u-boot SHA to use the latest patches.

These patches switch u-boot and the kernel to using UART6 (ttymxc5) on
the mikroBus for the console output.

To use the mikroBus UART, you can either add an RS232 click board [1],
or connect a TTL serial to USB converter, such as the one used for the
RPi3 [2], to the GND, TX and RX pins on the WaRP7 mikroBus connector.

Male-to-male header pins are needed to convert the WaRP7's female header
to the TTL cable's female header.

[1] https://www.mikroe.com/rs232-click
[2] http://www.ftdichip.com/Products/Cables/RPi.htm

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>