IOTMBL-140: rpi3 image doesn't boot (kernel panic)
The partition scheme implemented on RPi3 by
mbed_sdcard_image-rpi.bbclass didn't match what the U-Boot script
expected.

Getting things right with mbed_sdcard_image-rpi.bbclass is much harder
than just specifying the partition layout in a wks file and allowing WIC
to do the rest, so switch to using WIC for our RPi3 SD card images and
update the U-boot script to match.

The wks file added for RPi3 is the same as the wks file for Warp7 except
that there's no partition for a raw U-boot image and the boot partition
is a little bit bigger.

Add some comments to the RPi3 U-boot script because the U-boot commands
can be a little cryptic.

Also resolves:
IOTMBL-76: See if we can use WIC for creating rpi3 images

Signed-off-by: Jonathan Haigh <jonathan.haigh@arm.com>