u-boot: Update to drive booting via boot.scr
ARM has a clear desire to drive most of the boot logic via boot.scr and/or
boot.scr-signed respectively. This requires dropping a bunch of patches in
u-boot and moving the functionality into boot.scr.

Its actually a good opportunity to reduce down some patches and rethinking
some of the logic in the secure-boot chain to make it a bit easier to
understand.

These two checkins bring both the u-boot binary and the mkimage to the same
SHA. This version of u-boot is based off of v2018.03-rc1 which contains MBL
fixes that have been upstreamed to u-boot already specifically

- HAB fixes (lots of)
- CAAM fixes - to mitigate CAAM failure when booting in secure-boot mode.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>