u-boot: Remove optee specific config
A change in the 2018.03-rc2 branch broke the OPTEE defconfig. Looking at
the current differences between the secure_deconfig and
secure_optee_defconfig it looks like there's no good reason to have a
forked version of the secure defconfig just for OPTEE. This patch drops the
OPTEE specific defconfig in favour of the upstream warp7 secure config with
HAB enabled.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>