u-boot: include u-boot meta-data for offline code-signing
In order to support signing of images outside of yocto we need a few
specific files from the u-boot build.

- u-boot.bin

  To enable direct signing of the binary before rewriting it to the
  SD image.

- u-boot.cfg

  Which contains the load addresses which a given u-boot binary will
  load particular images to. Since the u-boot load address and the HAB
  load address must be the same - using u-boot.cfg as the canonical
  repository for this data makes sense.

- imximage.cfg.cfgtmp

  Contains the "CSF" definition used as an input to u-boot mkimage
  Without the CSF definition we will not generate the correct type
  of IVT header for the BootROM to parse.

With these three files provided inside the boot partition - we can

- Sign u-boot.bin
- Sign any arbitrary file provided we use a load address defined
  in u-boot.cfg

This patch modifies the boot partition produced to include all three
binaries.

Other patches in the mbl-tools repository make use of these files when
signing.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>