IOTMBL-201: Ensure linux-firmware_%.bbappend complies with relevant licences (#55)
Note: I've switched to calling license files "COPYING" rather than
"LICENSE" because "COPYING" rather than "LICENSE" seems to be the standard name for
license files used by OpenEmbedded layers.

recipes-kernel/linux-firmware/linux-firmware_%.bbappend:
This file was copied from the meta-raspberrypi OpenEmbedded layer
(https://github.com/agherzan/meta-raspberrypi/) and modified so it must
comply with the license for that file (MIT). That means the original
content should be under the MIT license and we should preserve copyright
information. No copyright information was available for this file so
none has been copied yet. I will request copyright information from the 
meta-raspberrypi maintainers so that we can add it in the future.

recipes-kernel/linux/linux-warp7_%.bbappend:
This file was also copied from elsewhere and is covered by the MIT
license. It already has a copyright notice and a license specifier, but
let's add an SPDX license specifier to make the repo's non-Apache-2.0
parts easier to track. We don't have the "COPYING.MIT" referenced
here so add a copy of the MIT license to the file.

LICENSE -> COPYING.Apache-2.0:
This repo contains content that is not under the Apache-2.0 license so
move the Apache-2.0 license into an Apache-2.0 specific file.

COPYING:
Note that the repo has content under different licenses and give rules
for determining which content is under which license.

Signed-off-by: Jonathan Haigh <jonathan.haigh@arm.com>