mbl-boot-src: imx7s-warp: set rootpart to mmcpart
In order not to break the existing upstream u-boot secure for WaRP7 we need
to differentiate the root-partition from the mmcpart - which are not the
same thing in the upstream.

We select the desired slot by way of boot.scr so plug the selected mmcpart
variable into the rootpart variable. Resolving UUID will then 'just work'
on the regular WaRP7 board, the WaRP7 secure boot and the MBED Linux WaRP7
secure boot.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>