u-boot: Use latest SHA for u-boot and mkimage
As a precursor to pushing out the final stage of secure-boot modification
to upstream we should sync-up to the latest version of u-boot which already
has some important changes from our project included - primarily the
IH_OS_TEE change enabling creation of a bootable OPTEE image.

This patch updates u-boot and u-boot-mkimage in lockstep.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>