IOTMBL-219: Add Ethernet Gadget driver
In a previous commit the kernel config was modified to build the g_ether
module into the kernel. There were issues with that - it was fragile due
to variations in the rest of the config that was coming from other
sources.

It turns out that there's a "usbinit" package that just does the right
thing to get usb0 up during boot, so use that instead of fiddling with
the kernel config.

classes/mbl_kernel_config.bbclass:
* This is no longer required.

recipes-kernel/linux/linux-raspberrypi_%.bbappend:
recipes-kernel/linux/linux-warp7_%.bbappend:
* Don't inherit the no longer required or existent mbl_kernel_config
  class.

recipes-core/packagegroups/packagegroup-mbl.bb:
* Include the usbinit package to deal with bringing up usb0 (but only when
  the machine supports being a USB gadget).

recipes-core/init-ifupdown/init-ifupdown/interfaces:
* usbinit now runs "ifup" for usb0 during boot so don't try to start it
  automatically with "auto usb0".

Signed-off-by: Jonathan Haigh <jonathan.haigh@arm.com>