pki/nxp/sign/Makefile: Switch to warp7_defconfig
In discussion with Fabio Estevam the upstream NXP maintainer for WaRP7
we've discovered that its possible to support both secure and non-secure
with what was previously called the secure defconfig.

As a result the secure and non-secure defconfigs have been unified into
warp7_defconfig.

This patch reflects the pending upstream change in the Makefile here.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>