pki/nxp/sign: Make it possible to specify SRK table name
The default SRK table name is SRK_1_2_3_4_table.bin but for various reasons
that may not be the name that is ultimately used by a user of this script
when signing.

This patch makes it possible to specify the SRK table name via the
environment.

If not specified the name defaults to the default name the CST tool will
choose.

SRK_1_2_3_4_table.bin

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>