pki/nxp/sign/Makefile: Fix copy of cst data to target directory
The current make target ${CRTS_DATA}/${SRK_FUSE}: will check if the source
files exist - which they do, and then fail to copy the source files to the
destination directory.

Ensure we have a make target that will copy the required files each time.

Note: over-writing the csf files on each make run is a good idea so that
subsequent make runs do not append to the csf files from a previous run.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>