Remove upstreamed patch from U-Boot recipe
openembedded-core updated the version of U-Boot it uses, and the new
version includes the changes that meta-raspberrypi's
0002-rpi_0_w-Add-configs-consistent-with-RpI3.patch makes, so that patch
from meta-raspberrypi is no longer necessary (and doesn't apply
cleanly).

We currently have a pinned version of meta-raspberrypi, so even if this
gets fixed in that repo we won't pick up the fix.

Add some code in meta-mbl to remove that patch from the U-Boot recipe so
that our builds will work until meta-raspberrypi is fixed and we unpin
our version of that repo.

Signed-off-by: Jonathan Haigh <jonathan.haigh@arm.com>