Revert "IOTMBL-372: Fix for incorrect linux-firmware_git.bb md5 checksums."
This reverts commit 24ef15e1b0cc77b9fd0ea23cbf8173f8df737622

The commit being reverted appears to be based on the Rocko branch. It
introduced changes to _remove LICENCE.Netronome with the wrong md5sum,
then reinstate it with the same value as already exists in the original
linux_firmware_git.bb recipe.

The change doesn't cause problems today, but if LICENCE.Netronome
changes, the build will break.

Commit cc928e6c10ee0f261af03b918ab464c4f40e5f2d "linux-firmware: Fix
LIC_FILES_CHKSUM" already reverted the changes to the WHENCE md5sum.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>