walkthrough: don't use alpha branch in master branch walkthrough
For:
IOTMBL-355: master branch walkthrough.md tells user to use alpha branch
code but expects bmap files to be generated

The walkthrough on the master branch should contain instructions for
building the master branch and instructions for building other
branches should be left to those branches themselves.

Make this walkthrough consistently refer to the master branch, and add a
comment to help the reader understand which document they should use.

Signed-off-by: Jonathan Haigh <jonathan.haigh@arm.com>