mbl-boot-scr: remove dtb and optee entries
With the use of ARM Trusted Firmware boot flow the DTB and optee are now handle
and validated by the FIP image and because of that not loaded by the u-boot as
BL33.
So, remove the entries related to DTB and optee from the boot script and move
the bootm command from loading optee to load the linux kernel.

Signed-off-by: Rui Miguel Silva <rui.silva@linaro.org>