IOTMBL-418 - initramfs prototype implementation
Add an init script (initramfs-init-script.sh), busybox and util-linux-findfs to the initramfs embedded in the Linux kernel image.

initramfs-init-script.sh selects rootfs partition based on the existence of rootfs file in the bootflags partition.
This is needed so that we can later do some dm-verity initialization and add signature verification for root-hash (needed by dm-verity).

initramfs image and kernel are combined together to form a single image as it is recommended to bundle the initramfs image with the kernel image to avoid
circular dependencies between the kernel recipe and the initramfs recipe should the initramfs image include kernel modules.