rpi3: support FIT image for u-boot
This commit enables the usage of U-boot FIT image. Thus the kernel and
dtb and bootscripts are packaged into FIT and can be verify by u-boot.

Migrates to using RPi3B+ instead of RPi3B.

We need to put FIT image into FAT partition. Previously it is 20MB. And
filled with VideoCore4 boot binarnies. We enlarge it to 40MB.

Signed-off-by: Ying-Chun Liu (PaulLiu) <paulliu@debian.org>
Signed-off-by: Jun Nie <jun.nie@linaro.org>