IOTMBL-579: Create ext4 image of rootfs partition. (#167)
The standalone ext4 image is required by veritysetup for calculating dm-verity’s hash tree of the rootfs so that later on we will be able to protect rootfs on the device using dm-verity.

Leave "tar.xz" in IMAGE_FSTYPES because that's still what our over-the-air firmware update process uses. Once the dm-verity work is complete and the over-the-air firmware update process has been updated to support it we can remove "tar.xz".