u-boot: Switch to using warp7_bl33_defconfig
In order to differentiate between WaRP7 run directly by the BootROM and
WaRP7 run by OPTEE we require a different defconfig and to skip over
some low-level initialization in u-boot.

Skipping the low-level initialization is a result of running u-boot in a
lower privilege level after OPTEE runs and hands over to u-boot.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>