mbl-cloud-client recipe adaptation for bbappend file (#180)
Now SRC_URI variable is a concatenation of the two parts. The change is caused by the fact, that patches, which are added in mbl-cloud-client.bb, do not apply on sources of the mbl-cloud-client.bbappend file (which located in meta-mbl-internal-extras repo). The first part of SRC_URI is the common (for mbl-cloud-client and mbl-cloud-client-internal) files. The second part is the specific sources(the git repository, and the patch files). The separation of SRC_URI into two parts make it easier to form SRC_URI correctly in the mbl-cloud-client.bbappend file.

Signed-off-by: MaximAxelrod <maxim.axelrod@arm.com>