recipes-kernel: war7: ATF: Do early installation of DTB
In order to make sure the DTB is installed in a location that can be
accessed by the ATF/FIP logic during the do_compile() phase of the ATF
recipe, we need to make the kernel DTB available early.

This patch does that by installing the DTB in a do_install_append()
routine.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>