recipes-kernel: warp7: FIT: Add Kernel FIT image generation to build
This patch adds a file to describe and a bitbake recipe to generate a Linux
kernel FIT image.

u-boot then can verify the kernel image prior to booting it. The FDT will
already have been placed in-memory by ATF and should have been verified
already, so, no additional validation of the in-memory DTB is performed.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>