IOTMBL-751 Add dummy OpenCV recipe to fix build error
Due to an upstream update to meta-openembedded, the OpenCV recipe the
meta-freescale layer was appending has been upversioned. To fix the MBL
build error, a dummy recipe matching the PV of the bbappend file is to be
added at layers/meta-mbl/recipes-support/opencv_3.3.bb.
This dummy recipe is to be removed as soon as meta-freescale is updated
to append the correct version of OpenCV in meta-openembedded.