Master+linaro warp7 atf+fit (#202)
* u-boot: Move u-boot to FIT image based boot.scr
Moving to the FIT format for boot.scr requires updating the u-boot
defconfig. This patch sets the appropriate SHA revision number in our
Linaro landing team git repo.

* mbl-boot-scr.bb: warp7: Convert to using boot.its
This is mostly a copy-and-paste from Paul Liu's work on the RPI3.

* recipes-kernel: warp7: FIT: Add Kernel FIT image generation to build
This patch adds a file to describe and a bitbake recipe to generate a Linux
kernel FIT image. u-boot then can verify the kernel image prior to booting it. The FDT will
already have been placed in-memory by ATF and should have been verified
already, so, no additional validation of the in-memory DTB is performed.

* mbl-boot-src: imx7s-warp: Switch to loading kernel itb
A kernel.itb file will describe the load-location and signing specifics of
the Linux kernel and initramfs. We need to modify boot.cmd to load the
kernel.itb to a known location and then boot.

* imx7s-warp: atf: Tidy up build collateral
This patch tidies up some of the build collateral hanging around in the
deploy directory.