Master+linaro warp7+atf tbb v2 (#252)
* atf: atf-fiptool-native: Rename recipe to atf-tools-native

Later patches will make use of other tools provided by the ATF tools
directory as a result the name atf-fiptools_git.bb will no longer be
accurate.

Change the name now to allow subsequent patches be atomic in their addition
of new tools.

* atf: atf-fiptools-native: Add cert_create to FIP tools

This patch adds cert_create to ATF tools. certe_create is required to
generate certificate data when signing binaries in a FIP.

* atf: warp7: Convert to ATF Trusted Board Boot

This patch change the ATF WaRP7 recipe to build ATF for Trusted Board Boot.
Once done the out-of-the-box ATF code for establishing a root-of-trust from
BL2 is enabled.

Used in conjunction with NXP's own SoC verifiction of BL2 - High Assurance
Boot (HAB) it is possible to have a root-of-trust that is compliant with
ATF's TBB flow.