Fix incremental build issue in image-signing.bbclass (#336)
For: IOTMBL-1025: [meta-mbl#316] atf-warp7 recipe seems to break Warp7
Incremental Builds

The do_sign_root_hash task uses an artefact (the "root hash") left by
the do_generate_verity_metadata task (image-verity.bbclass), but the
name of that artefact has a timestamp in it (from $IMAGE_NAME). When
do_sign_root_hash runs, $IMAGE_NAME may have a different timestamp to
when do_generate_verity_metadata ran, and in that case do_sign_root_hash
fails to find the artefact.

Fortunately, after do_generate_verity_metadata creates the root hash, it
also creates a symlink to it with a name that doesn't contain the
timestamp, so use that symlink in do_sign_root_hash to avoid the
problem.