atf: rationalize uses of "+=" and "_append"
In atf-raspberrypi3-mbl_git.bb there was:
DEPENDS_append += " linaro-aarch64-toolchain-native "
This is confused - only one of "+=" and "_append" should be used for a
single string operation.

In this and other atf .bb files "_append" and "+=" were both used to
append to variables with no apparent reason why one was chosen over the
other.  In oe-core, "+=" seems to be the default when either would do,
and "_append" is only used when it is required.  Let's follow that
convention and just stick to "+=" in these files.

Jira issue: IOTMBL-1025
GitHub issue: #316