imx-boot: Fix name of DTB with compile_mx8_prepend()
The do_compile() phase of the incoming imx-boot_0.2 recipe from NXP expects
to find a file named u-boot-nodtb.bin in a subdirectory of the deploy
directory.

In our case the file is named u-boot-nodtb.bin-${MACHINE}-${UBOOT_CONFIG}.
This patch fixes the discongruity by symlinking the compiled file to the
expected name.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>