dbus: Install init.d scripts for both sysvinit and systemd images
On this first round of work for having systemd as a replacement of
the sysvinit we are going to use the init.d scripts as the service
units files. Later on we will create the appropriated services files.

This commit also removes trailing spaces in the recipe.

Part of IOTMBL-1523: SysVinit->Systemd: get developement image with
boot prompt