Add firmware update header in disk images (take 2) (#377)
For IOTMBL-1099 [https://github.com/ARMmbed/mbl-core/issues/46]: Version
of MBL is not reported correctly in Pelion Cloud

When mbl-cloud-client wants to know what version of firmware is running
on a device it runs a script "arm_update_active_details.sh" that passes
mbl-cloud-client a "HEADER" file that contains information about the
last firmware update.

Usually, this HEADER file is passed to the device along with a firmware
update from the Pelion cloud, but this doesn't happen when flashing the
first full disk image to the device so we must generate a HEADER file
and add it to the disk image.

Ideally, we'd follow the example of image-buildinfo and just add a
function that creates the HEADER file to IMAGE_PREPROCESS_COMMAND.
Unfortunately, IMAGE_PREPROCESS_COMMAND functions are run using the host
Python installation rather than the "native" python installation built
by bitbake. We need to use the "native" python installation because we
need access to the mbl-firmware-update-header-util library which is
probably not available in the host Python installation.

Therefore, put the python code that creates the HEADER file into a
script. We can execute the script using the "native" python installation
from an IMAGE_PREPROCESS_COMMAND function.

We actually need to add this HEADER file to the bootflags partition, so
update our wks files to populate this partition rather than leaving it
empty as we do now.