Fix breakage due to upstream python3 recipe changes (#380)
For IOTMBL-1621: Build breakage due to upstream python3 recipe change

The python3 recipe in oe-core has changed and it doesn't play nicely
with our python3 bbappend which does some pretty messy things to allow
us to avoid including gdbm (which leads to GPLv3 in the image).

We've decided that we're not going to be able to avoid GPLv3 on our root
file system partition, so we can get rid of the nastyness involving gdbm
in python3.

So:
* Get rid of the database lib handling in our python3 bbappend.
* Remove our python3 dbm related PACKAGECONFIG change in mbl.conf.
* Change our GPLv3 checking code to use different license blacklists for
  packages installed on the rootfs and recipes that contribute to the
  non-rootfs parts of the image - we still don't want GPLv3 in e.g.
  bootloaders.
* Update our image recipes to allow GPLv3 on the root file system but
  not in the rest of the disk image.
* Update our initramfs image recipe to disallow GPLv3 entirely - we
  should never have been allowing it.

Additionally, we now seem to require the python3-pickle package in order to
successfully run "python3 -m pip install -U pytest". This might be an issue
with python3-pip's RDEPENDS, but just add it to our package group to fix the
build for now.

Pytest now complains about the "pathlib" module being missing. There is no
python3-pathlib package defined by the upstream python3 recipe, but there is a
pathlib library in the python3 source code so add a python3-pathlib package.