mbl-app-lifecycle-manager: Remove init shell script (#403)
For IOTMBL-1762 mbl-app-lifecycle-manager: get rid of
mbl-app-lifecycle-manager-init.sh

The shell script does not have enough content to justify the existence
of the shell script. The content of the shell script is instead added to
the application systemd service file.
This commit also makes use of the `SRCNAME` BitBake variable to form
the systemd service file name in the package recipe. This is done to
ensure a single point of modification in case of renaming the
package