mbl-dbus-cloud: Add systemd socket and rework service file
Following the implementation for the main d-bus system-wide bus service
file, we added a socket file and make the service file requiring it.

Also, in the service file add the --nopidfile option and remove this
configuration from mbl-dbus-cloud.conf because systemd is capable of
managing this service without a pid file.

Part of IOTMBL-1695: Systemd PR370 follow-on tasks