systemd-net-conf: systemd-net-conf-dbg-iface: Rework debug interface
Rename the recipe from "systemd-network-conf" to "systemd-net-conf" in
order to use a shorter name and remove the debug interface package
generation. This recipe now only provides a single package setup the
hostname.

Get rid of the "ethernet" MACHINE_FEATURE and make usage of
MBL_DEBUG_INTERFACE variable set in the machine conf file. The
systemd-net-conf-dbg-iface recipe provides the network configuration
file for the debug interface to be installed in development image.
The 10-debug-interface.network has the
__REPLACE_ME_WITH_MBL_DEBUG_INTERFACE__ placeholder to be replaced by
the MBL_DEBUG_INTERFACE value. If this value is not set we raise a
bitbake build error.

Part of IOTMBL-1695: Systemd PR370 follow-on tasks