Network interfaces configuration rework
It was found that in some cases the network interfaces names are
different from what was set during the build time. This can happen
if the order of the kernel modules load is different across reboots,
for example.

Because of this we decided to remove the MBL_DEBUG_INTERFACE variable
and just use systemd to set the link local address in all interfaces
and defer to connman the management of those interfaces, including the
debug and internet connections.

Part of IOTMBL-1823: Conflict between cellular module and usb-gadget