optee-os: Align imx7s, imx7d and rpi3 to same SHA
All of the necessary changes to OP-TEE for the imx7s, imx7d and rpi3 are
upstream. This patch is part of a set of patches which is addressing an
error with xtest where a test fails on an invalid input parameter.

The root cause is misalignment between xtest and optee-os. For simplicity
this patch brings the two imx7 boards and the rpi3 to the same SHA.

It is not yet possible to include the imx8 in that due to the upstreaming
status of the imx8 but it too will be included in the one SHA at a later
date.

For now align just the imx7s, imx7d and rpi3.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>