connman: Add specific raspberrypi3-mbl config for debug interface
Because we configure connman to give priority to ethernet over wifi
(PreferredTechnologies=ethernet,wifi), for the raspberrypi if we have
the USB to ethernet adapter plugged and connected the wifi will not
connect automatically, unless the user issues a command to connect.

This patch adds NetworkInterfaceBlacklist=eth1 config for connman
assuming that the "eth1" is the debug interface.

Fixes: IOTMBL-1903: Rpi3: cannot use/test wifi when usb-dongle is
connected