mbl-usb-gether-mac-addr: Persist usb gether MAC addr across updates
We have to persist the usb gether MAC addr across reboots and the
current implementation relies on saving these information in
/etc/modprobe.d/g_ether.conf . This path doesn't persist after an
update and we can't symlink it to /config/user/modprobe.d/g_ether.conf
because in the early stages of the boot process (before running
systemd-modules-load.service) the /config is not mounted.

To overcome this restriction we are removing the g_ether from the
KERNEL_MODULE_AUTOLOAD and loading the g_ether module with the MAC addr
in the mbl-usb-gether-mac-addr.service.

Fixes #470