Remove layers and recipes not required for rpi3 secure boot support
For IOTMBL-1891: Implement Yocto community branching in meta-mbl-layers

This branch will be limited to providing secure boot on raspberrypi3
to the core-os/gateway project so get rid of layers and recipes not
required for that. This includes removing linux kernel changes because
the core-os/gateway will continue using their own kernel.