Move INITRAMFS_IMAGE definition to distro conf (#484)
The choice of initramfs image (or even whether to use an initramfs) is a
distro choice rather than machine config so move the definition of
INITRAMFS_IMAGE to mbl.conf.

INITRAMFS_IMAGE used to be specified in machine conf files because when
adding support for a new board we are sometimes in a state of partial
support where the board does not yet use MBL's initramfs. It is
convenient in these cases to specify INITRAMFS_IMAGE in machine conf
files.

Now, however, we need to support users who will use our BSP layers but
not our distro layer. In this case, our BSP users really shouldn't get
our INITRAMFS_IMAGE setting so we'll just have to deal with partial
board support differently in the future.