Move WKS_FILE definitions to distro conf (#489)
For IOTMBL-1891: Implement Yocto community branching in meta-mbl-layers

The .wks file to use for a build is machine dependent, but also distro
dependent. MBL's .wks files would not be suitable for users who take
MBL's BSP but not the full MBL distro so move the WKS_FILE definitions
out of our machine configs and into our distro config.