imx8mm_evk_mbl: atf: Switch on an initial empty FIP boot
This commit enables bundling of an empty FIP with the target ATF image.
Once done the ATF binary will open but, not extract any elements from the
FIP, since there is nothing in the FIP right now to be extracted.

The DEBUG_CONSOLE is switched on to make it easier to see ATF booting and
doing something useful.

Due to the older nature of the TFA sources we need to have some additional
commits backported to the TFA codebase; added to the Linaro landing team
repo.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>