meta-virtualization-mbl:cgroup-lite: Fix SYSTEMD_AUTO_ENABLE_${PN}
With the commit 154abbc3296eded11d2bbe3e102470b6986d42cd in
openembedded-core the SYSTEMD_AUTO_ENABLE needs to be set to 'enable'
or 'disable'.

The cgroup-lite recipe from meta-virtualization sets it to 'mask' which
is not allowed anymore.

We introduce this bbappend to have
SYSTEMD_AUTO_ENABLE_${PN} = "disable".