imx8mm-evk: fip: Fix unnoticed build error
Commit 7f3ef3f63f2a ("Bod warrior dev+optee fip precursor+imx8 empty fip
(#527)") has an error with-respect to the location of the unified BL1/FIP
binary, which due to an unlikely set of events on Jenkins, was not noticed
until after merging.

This patch fixes the original error by removing the do_deploy() logic from
the script temporarily, we won't need to mess with do_deploy() until
switching on FIP verification.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>