layers: apply new naming convention for 3rd party layers customization
This patch applies the new convention for naming 3rd party layers
customization directories. Now the "-mbl" string should be appended
only in the repository directory name and not in the meta layer
directory. Some cases the repository directory is the meta layer
directory.

Other change introduced is that the BBFILE_COLLECTIONS should append
"-mbl" to the ustream layer original value.