imx8mm_evk_mbl: atf: Switch to BL2/FIP chainloading
This commit switches our i.MX8 branch to a branch that chainloads BL2 ->
FIP -> {BL31, BL32, BL33} with FIP signing enabled.

We need to change the name of the intermediary binary from bl31-something
to bl2-something, since, we are using BL2 not BL31.

Other than that most of the work is contained in the ATF branch itself.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>